# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Baseball Matching","description":"野球チームのマッチングアプリケーション","version":"1.0"},"host":"localhost:8080","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/v1/users":{"get":{"tags":["User"],"summary":"List User User","description":"ユーザーの一覧を取得します。\n\nRequired security scopes:\n","operationId":"User#List User","parameters":[{"name":"Authorization","in":"header","description":"JWT used for Authentication","required":false,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserBbmatchingUserResponseCollection"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/UserListUserUnauthorizedResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]},"put":{"tags":["User"],"summary":"Update current user User","description":"現在のセッションに紐づくユーザーの情報を更新します。\n\nRequired security scopes:\n","operationId":"User#Update current user","produces":["application/json; charset=utf-8"],"parameters":[{"name":"Authorization","in":"header","description":"JWT used for Authentication","required":false,"type":"string"},{"name":"Update Current UserRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserUpdateCurrentUserRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserUpdate Current UserResponseBody"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/UserUpdateCurrentUserUnauthorizedResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]},"delete":{"tags":["User"],"summary":"Delete current user User","description":"現在のセッションに紐づくユーザーを削除します。\n\nRequired security scopes:\n","operationId":"User#Delete current user","parameters":[{"name":"Authorization","in":"header","description":"JWT used for Authentication","required":false,"type":"string"}],"responses":{"204":{"description":"No Content response."},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/UserDeleteCurrentUserUnauthorizedResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/v1/users/jwt/{user_id}":{"get":{"tags":["User"],"summary":"Get JWT User","description":"指定したユーザーIDのJWTを取得します","operationId":"User#Get JWT","parameters":[{"name":"user_id","in":"path","description":"firebaseのユーザーID","required":true,"type":"string","maxLength":28,"minLength":28}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserGet JWTResponseBody"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/UserGetJWTUnauthorizedResponseBody"}}},"schemes":["http"]}},"/v1/users/me":{"get":{"tags":["User"],"summary":"Get current user User","description":"現在のエンドポイントに紐づくユーザーの情報を返します。\n\nRequired security scopes:\n","operationId":"User#Get current user","produces":["application/json; charset=utf-8"],"parameters":[{"name":"Authorization","in":"header","description":"JWT used for Authentication","required":false,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserGet Current UserResponseBody"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/UserGetCurrentUserUnauthorizedResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/v1/users/{user_id}":{"get":{"tags":["User"],"summary":"Get User User","description":"指定したIDのユーザーの情報を取得します。\n\nRequired security scopes:\n","operationId":"User#Get User","produces":["application/json; charset=utf-8"],"parameters":[{"name":"user_id","in":"path","description":"firebaseのユーザーID","required":true,"type":"string","maxLength":28,"minLength":28},{"name":"Authorization","in":"header","description":"JWT used for Authentication","required":false,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserGet UserResponseBody"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/UserGetUserUnauthorizedResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}}},"definitions":{"BbmatchingUserResponse":{"title":"Mediatype identifier: application/vnd.bbmatching.user+json; view=default","type":"object","properties":{"UserName":{"type":"string","description":"チームの表示名","example":"Chunichi Dragons"},"email":{"type":"string","description":"チームのプライマリメールアドレス","example":"chunichi@example.com","format":"email"},"email_verified":{"type":"boolean","description":"ユーザーのプライマリメールアドレスが確認されているか。","example":false},"phoneNumber":{"type":"string","description":"チームのメイン電話番号","example":"09012345678","pattern":"^\\+?[\\d]{10,}$"},"photoURL":{"type":"string","description":"チームの写真URL","example":"http://img.com"},"user_id":{"type":"string","description":"firebaseのユーザーID","example":"XRQ85mtXnINISH25zfM0m5RlC6L2","minLength":28,"maxLength":28}},"description":"User Response (default view)","example":{"UserName":"Chunichi Dragons","email":"chunichi@example.com","email_verified":false,"phoneNumber":"09012345678","photoURL":"http://img.com","user_id":"XRQ85mtXnINISH25zfM0m5RlC6L2"},"required":["user_id","UserName","email","phoneNumber","photoURL","email_verified"]},"UserBbmatchingUserResponseCollection":{"title":"Mediatype identifier: application/vnd.bbmatching.user+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/BbmatchingUserResponse"},"description":"List UserResponseBody is the result type for an array of BbmatchingUserResponse (default view)","example":[{"UserName":"Chunichi Dragons","email":"chunichi@example.com","email_verified":true,"phoneNumber":"09012345678","photoURL":"http://img.com","user_id":"XRQ85mtXnINISH25zfM0m5RlC6L2"},{"UserName":"Chunichi Dragons","email":"chunichi@example.com","email_verified":true,"phoneNumber":"09012345678","photoURL":"http://img.com","user_id":"XRQ85mtXnINISH25zfM0m5RlC6L2"}]},"UserDeleteCurrentUserUnauthorizedResponseBody":{"title":"UserDeleteCurrentUserUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Eaque eligendi."},"UserGet Current UserResponseBody":{"title":"Mediatype identifier: application/vnd.bbmatching.user+json; view=default","type":"object","properties":{"UserName":{"type":"string","description":"チームの表示名","example":"Chunichi Dragons"},"email":{"type":"string","description":"チームのプライマリメールアドレス","example":"chunichi@example.com","format":"email"},"email_verified":{"type":"boolean","description":"ユーザーのプライマリメールアドレスが確認されているか。","example":false},"phoneNumber":{"type":"string","description":"チームのメイン電話番号","example":"09012345678","pattern":"^\\+?[\\d]{10,}$"},"photoURL":{"type":"string","description":"チームの写真URL","example":"http://img.com"},"user_id":{"type":"string","description":"firebaseのユーザーID","example":"XRQ85mtXnINISH25zfM0m5RlC6L2","minLength":28,"maxLength":28}},"description":"Get Current UserResponseBody result type (default view)","example":{"UserName":"Chunichi Dragons","email":"chunichi@example.com","email_verified":true,"phoneNumber":"09012345678","photoURL":"http://img.com","user_id":"XRQ85mtXnINISH25zfM0m5RlC6L2"},"required":["user_id","UserName","email","phoneNumber","photoURL","email_verified"]},"UserGet JWTResponseBody":{"title":"Mediatype identifier: application/vnd.bbmatching.jwt+json; view=default","type":"object","properties":{"JWT":{"type":"string","description":"Json Web Token","example":"eyJhbGciOiJSUzI1NiJ9.eyJzdWIiOiJ1c2VyaG9nZSIsImF1ZCI6ImF1ZGhvZ2UiLCJpc3MiOiJodHRwczpcL1wvZXhhbXBsZS5jb21cLyIsImV4cCI6MTQ1MjU2NTYyOCwiaWF0IjoxNDUyNTY1NTY4fQ.BfW2a1SMY1a8cjb7ATcJPwYSmB1P6l4oN2QRNtod-xCyochsB1ZurxNLqPOGFr7_Abqpk-_lOUaPOdL2jQ23T1DS1lmQgaEZgUXaJPAqGSJygANpv8ds07Q6pbX_XbFpJdoVCQHzP8MjbjW_ft2ZAAJzjZZEC6Hm0WUxKS6V0yRNmEUyV-Ljh6n337rtAoSTks2APdgW1hffyZMrptKKazG2m0V0LQRnu5lmWLTYwTZC8NdNJZNepPwQiGNZS0IcrdZhguGAL75ZGTMw9O_EC9gv_I_9I5NUwZk6LG1feEy3MawT0QaTEsF5n6yUKJ8ziuMXnUEsymGdKC-VYEbPyw"}},"description":"Get JWTResponseBody result type (default view)","example":{"JWT":"eyJhbGciOiJSUzI1NiJ9.eyJzdWIiOiJ1c2VyaG9nZSIsImF1ZCI6ImF1ZGhvZ2UiLCJpc3MiOiJodHRwczpcL1wvZXhhbXBsZS5jb21cLyIsImV4cCI6MTQ1MjU2NTYyOCwiaWF0IjoxNDUyNTY1NTY4fQ.BfW2a1SMY1a8cjb7ATcJPwYSmB1P6l4oN2QRNtod-xCyochsB1ZurxNLqPOGFr7_Abqpk-_lOUaPOdL2jQ23T1DS1lmQgaEZgUXaJPAqGSJygANpv8ds07Q6pbX_XbFpJdoVCQHzP8MjbjW_ft2ZAAJzjZZEC6Hm0WUxKS6V0yRNmEUyV-Ljh6n337rtAoSTks2APdgW1hffyZMrptKKazG2m0V0LQRnu5lmWLTYwTZC8NdNJZNepPwQiGNZS0IcrdZhguGAL75ZGTMw9O_EC9gv_I_9I5NUwZk6LG1feEy3MawT0QaTEsF5n6yUKJ8ziuMXnUEsymGdKC-VYEbPyw"}},"UserGet UserResponseBody":{"title":"Mediatype identifier: application/vnd.bbmatching.user+json; view=default","type":"object","properties":{"UserName":{"type":"string","description":"チームの表示名","example":"Chunichi Dragons"},"email":{"type":"string","description":"チームのプライマリメールアドレス","example":"chunichi@example.com","format":"email"},"email_verified":{"type":"boolean","description":"ユーザーのプライマリメールアドレスが確認されているか。","example":true},"phoneNumber":{"type":"string","description":"チームのメイン電話番号","example":"09012345678","pattern":"^\\+?[\\d]{10,}$"},"photoURL":{"type":"string","description":"チームの写真URL","example":"http://img.com"},"user_id":{"type":"string","description":"firebaseのユーザーID","example":"XRQ85mtXnINISH25zfM0m5RlC6L2","minLength":28,"maxLength":28}},"description":"Get UserResponseBody result type (default view)","example":{"UserName":"Chunichi Dragons","email":"chunichi@example.com","email_verified":false,"phoneNumber":"09012345678","photoURL":"http://img.com","user_id":"XRQ85mtXnINISH25zfM0m5RlC6L2"},"required":["user_id","UserName","email","phoneNumber","photoURL","email_verified"]},"UserGetCurrentUserUnauthorizedResponseBody":{"title":"UserGetCurrentUserUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Non error omnis aliquid iste earum est."},"UserGetJWTUnauthorizedResponseBody":{"title":"UserGetJWTUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Ut saepe aut porro velit est nobis."},"UserGetUserUnauthorizedResponseBody":{"title":"UserGetUserUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Ad non quia illo ut."},"UserListUserUnauthorizedResponseBody":{"title":"UserListUserUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Eos quia saepe."},"UserUpdate Current UserResponseBody":{"title":"Mediatype identifier: application/vnd.bbmatching.user+json; view=default","type":"object","properties":{"UserName":{"type":"string","description":"チームの表示名","example":"Chunichi Dragons"},"email":{"type":"string","description":"チームのプライマリメールアドレス","example":"chunichi@example.com","format":"email"},"email_verified":{"type":"boolean","description":"ユーザーのプライマリメールアドレスが確認されているか。","example":true},"phoneNumber":{"type":"string","description":"チームのメイン電話番号","example":"09012345678","pattern":"^\\+?[\\d]{10,}$"},"photoURL":{"type":"string","description":"チームの写真URL","example":"http://img.com"},"user_id":{"type":"string","description":"firebaseのユーザーID","example":"XRQ85mtXnINISH25zfM0m5RlC6L2","minLength":28,"maxLength":28}},"description":"Update Current UserResponseBody result type (default view)","example":{"UserName":"Chunichi Dragons","email":"chunichi@example.com","email_verified":true,"phoneNumber":"09012345678","photoURL":"http://img.com","user_id":"XRQ85mtXnINISH25zfM0m5RlC6L2"},"required":["user_id","UserName","email","phoneNumber","photoURL","email_verified"]},"UserUpdateCurrentUserRequestBody":{"title":"UserUpdateCurrentUserRequestBody","type":"object","properties":{"UserName":{"type":"string","description":"チームの表示名","example":"Chunichi Dragons"},"email":{"type":"string","description":"チームのプライマリメールアドレス","example":"chunichi@example.com","format":"email"},"phoneNumber":{"type":"string","description":"チームのメイン電話番号","example":"09012345678","pattern":"^\\+?[\\d]{10,}$"},"photoURL":{"type":"string","description":"チームの写真URL","example":"http://img.com"}},"example":{"UserName":"Chunichi Dragons","email":"chunichi@example.com","phoneNumber":"09012345678","photoURL":"http://img.com"}},"UserUpdateCurrentUserUnauthorizedResponseBody":{"title":"UserUpdateCurrentUserUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Quis iusto cumque aperiam."}},"securityDefinitions":{"jwt_header_Authorization":{"type":"apiKey","description":"Secures endpoint by requiring a valid JWT token retrieved via the firebase.\n**Security Scopes**:\n","name":"Authorization","in":"header"}}}